update release process
I don't think there's any reason why version bumps need to be in a PR, it just slows down the process. This updates the last step to advise to just push version bumps to `main`.

Also, I think it's easier to create the tag at the same time as the release is drafted and it should still trigger the same release workflow. If this is not the case, I'm happy to adjust. In that case, I think it would make sense to write a script that automates the process.